Add seed data files for demo / testing
- seed_data.sql: bootstrap rows + 10 classes (5 Morning / 5 Day) +
  30 sections + 1500 students + 1500 enrollments
- subjects_seed.sql: 27 Bangladeshi NCTB-style subjects + class-subject
  mappings (Primary, Junior, SSC Science/Commerce/Humanities streams,
  1st/2nd papers, 4th-subject options)
- routine_exam_seed.sql: 2 exam routines, ~900 class_routine slots,
  ~160 exam_session rows, ECRA. Shift-aware times — pulls each class's
  shift_id and computes timings per shift (Morning 07:00 first bell /
  08:00–11:00 exams; Day 12:30 first bell / 13:00–16:00 exams).
  CTE pulls from class table directly, so adapts to any class count.
- accounts_schema.sql: fee_pricing.class_id BIGINT → INT to match
  class.id (Integer entity) — the original BIGINT was rejected by the
  FK constraint, causing the CREATE TABLE to silently fail.

+++ sql/accounts_schema.sql
@@ -68,7 +68,7 @@ CREATE TABLE IF NOT EXISTS fee_categories (
 CREATE TABLE IF NOT EXISTS fee_pricing (
   id                BIGINT          NOT NULL AUTO_INCREMENT,
   fee_category_id   BIGINT          NOT NULL,
-  class_id          BIGINT          NOT NULL,
+  class_id          INT          NOT NULL,
   amount            DECIMAL(15,2)   NOT NULL,
   effective_from    DATE                NULL,
   is_active         BIT(1)          NOT NULL DEFAULT b'1',
